fix: vote route logic (wip)

+++ app/api/discussions/[id]/vote/route.ts
@@ -18,43 +18,37 @@ export const PUT = async (req, { params }) => {
   const { vote } = await req.json();
 
   try {
-    const discussion = await Discussion.findById(id);
-
     const existingVote = await Vote.find({
       $and: [{ discussion_id: id }, { voter_id: session?.user?.id }],
     });
 
-    // Update existing vote
-    if (existingVote[0]) {
-      // Un-apply same vote
-      if (existingVote[0].type === vote) {
-        await Vote.deleteOne({ _id: existingVote[0]._id });
-        discussion.votes.pull(existingVote[0]._id);
+    if (!!existingVote[0] && existingVote[0].type === vote) {
+      await Vote.deleteOne({ _id: existingVote[0]._id });
 
-        if (vote === "up") {
-          discussion.upvotes--;
-        }
+      await Discussion.findByIdAndUpdate(
+        id,
+        {
+          $inc: {
+            [vote === "up" ? "upvotes" : "downvotes"]: -1,
+          },
+        },
+        { new: true }
+      );
+    } else if (!!existingVote[0] && existingVote[0].type !== vote) {
+      await Vote.updateOne({ _id: existingVote[0]._id }, { type: vote });
 
-        if (vote === "down") {
-          discussion.downvotes--;
-        }
-      }
-
-      if (existingVote[0].type !== vote) {
-        await Vote.updateOne({ _id: existingVote[0]._id }, { type: vote });
-
-        if (vote === "up") {
-          discussion.upvotes++;
-          discussion.downvotes--;
-        }
-        if (vote === "down") {
-          discussion.downvotes++;
-          discussion.upvotes--;
-        }
-      }
+      await Discussion.findByIdAndUpdate(
+        id,
+        {
+          $inc: {
+            [vote === "up" ? "upvotes" : "downvotes"]: 1,
+            [existingVote[0].type === "up" ? "upvotes" : "downvotes"]: -1,
+          },
+        },
+        { new: true }
+      );
     } else {
-      // Create new vote
-      const newVote = new Vote({
+      const newVote = await new Vote({
         discussion_id: id,
         voter_id: session.user.id,
         voter_name: session.user.name,
@@ -63,21 +57,20 @@ export const PUT = async (req, { params }) => {
 
       await newVote.save();
 
-      // Add new vote ref to discussion votes array
-      discussion.votes.push(newVote._id);
-
-      // Increment count accordingly
-      if (vote === "up") {
-        discussion.upvotes++;
-      }
-      if (vote === "down") {
-        discussion.downvotes++;
-      }
+      await Discussion.findByIdAndUpdate(
+        id,
+        {
+          $inc: {
+            [vote === "up" ? "upvotes" : "downvotes"]: 1,
+          },
+        },
+        { new: true }
+      );
     }
 
-    discussion.save();
+    const discussion = await Discussion.findById(id);
 
-    const jsonDoc = discussion.toJSON();
+    const jsonDoc = await discussion.toJSON();
 
     return NextResponse.json(
       {
